Unify  implementations and deprecate @lexical/selection
